feat: Lv4 로그인 기능 및 세션 인증 구현

+++ src/main/java/com/example/scheduledevelop/user/service/UserService.java
@@ -1,6 +1,6 @@
 package com.example.scheduledevelop.user.service;
 
-import com.example.scheduledevelop.schedule.entity.Schedule;
+import com.example.scheduledevelop.basic.SessionValue;
 import com.example.scheduledevelop.user.dto.*;
 import com.example.scheduledevelop.user.entity.User;
 import com.example.scheduledevelop.user.repository.UserRepository;
@@ -53,23 +53,31 @@ public class UserService {
 
     // 수정
     @Transactional
-    public UserUpdateResponse updateById(Long userId, UserUpdateRequest request) {
+    public UserUpdateResponse updateById(Long userId, UserUpdateRequest request, SessionValue sessionValue) {
+
         User updatedUser = userRepository.findById(userId).orElseThrow(
                 () -> new IllegalStateException("존재하지 않는 유저입니다.")
         );
+
+        if(!sessionValue.getUserId().equals(userId)){
+            throw new IllegalStateException("본인 정보만 수정 가능합니다.");
+        }
 
         updatedUser.update(request);
         return UserUpdateResponse.from(updatedUser);
     }
 
     // 삭제
-    public void delete(Long userId) {
+    public void delete(Long userId, SessionValue sessionValue) {
+
         User deletedUser = userRepository.findById(userId).orElseThrow(
                 () -> new IllegalStateException("존재하지 않는 유저입니다.")
         );
+        if(!sessionValue.getUserId().equals(userId)){
+            throw new IllegalStateException("본인만 삭제 가능합니다.");
+        }
         userRepository.delete(deletedUser);
 
     }
-
 
 }
